requirements.txt
fastapi 
uvicorn 
kafka-python 
scikit-learn 
pandas 
numpy 
joblib 
psycopg2-binary